+++ config/config_dev.go
@@ -4,9 +4,7 @@ package config
 
 var (
 	ChainId        = "gitopia"
-	GRPCHost       = "grpc.devnet.gitopia.com:9090"
 	GitServerHost  = "https://server.devnet.gitopia.com"
-	TmAddr         = "https://rpc.devnet.gitopia.com:443"
 	GasPrices      = "0.001ulore"
 	FeeGranterAddr = ""
 	Denom          = "ulore"