+++ config/config_testing.go
@@ -3,11 +3,9 @@
 package config
 
 var (
-	ChainId        = "gitopia"
-	GRPCHost       = "localhost:9090"
+	ChainId        = "localgitopia"
 	GitServerHost  = "http://localhost:5001"
-	TmAddr         = "http://localhost:26657"
 	GasPrices      = "0.001ulore"
-	FeeGranterAddr = ""
+	FeeGranterAddr = "gitopia12sjhqc3rqgvu3zpg8ekmwl005rp4ys58ekqg89"
 	Denom          = "ulore"
 )